correct discount type

+++ V8AddInLibrary/Protocols/BaseProtocol.cs
@@ -1233,7 +1233,7 @@ namespace CentralLib.Protocols
         /// <param name="value">% или сумма скидки/наценки, если процент то 9999(где 99.99% является макс значением)</param>
         /// <param name="comment">пояснительная строка</param>
         /// <returns></returns>
-        public virtual DiscountInfo Discount(FPDiscount typeDiscount, Int16 value, string comment)
+        public virtual DiscountInfo Discount(FPDiscount typeDiscount, int value, string comment)
         {
             logger.Trace(this.GetType().FullName + "." + System.Reflection.MethodBase.GetCurrentMethod().Name);
             //TODO test discount
